refactor camera manager and move image compression logic to viewmodel
