--- core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/camera/CameraManager.kt
@@ -1,11 +1,0 @@
-package net.thechance.mena.core_chat.presentation.camera
-
-import androidx.compose.runtime.Composable
-
-@Composable
-expect fun rememberCameraManager(onResult: (ByteArray?) -> Unit): CameraManager
-
-
-expect class CameraManager{
-    fun launch()
-}
