+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/chat/ChatScreen.kt
@@ -21,7 +21,6 @@ import dev.icerock.moko.permissions.compose.BindEffect
 import dev.icerock.moko.permissions.compose.rememberPermissionsControllerFactory
 import mena.core_chat_presentation.generated.resources.Res
 import mena.core_chat_presentation.generated.resources.you
-import net.thechance.mena.core_chat.presentation.camera.rememberCameraManager
 import net.thechance.mena.core_chat.presentation.screen.chat.components.AttachmentsBottomSheet
 import net.thechance.mena.core_chat.presentation.screen.chat.components.ChatHeader
 import net.thechance.mena.core_chat.presentation.screen.chat.components.ChatInputBar
@@ -29,6 +28,7 @@ import net.thechance.mena.core_chat.presentation.screen.chat.components.ChatList
 import net.thechance.mena.core_chat.presentation.screen.chat.components.ChatScreenOverlays
 import net.thechance.mena.core_chat.presentation.screen.chat.components.FullImagePagerView
 import net.thechance.mena.core_chat.presentation.utils.PaginationTrigger
+import net.thechance.mena.core_chat.presentation.utils.rememberCameraManager
 import net.thechance.mena.designsystem.presentation.component.scaffold.Scaffold
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import org.jetbrains.compose.resources.stringResource
@@ -61,12 +61,7 @@ fun ChatScreenContent(
     val chatListState = rememberLazyListState()
 
     val cameraManager = rememberCameraManager(
-        onResult = { byteArray  ->
-            byteArray?.let {
-                interactions.onSendImageClicked(listOf(byteArray))
-            }
-            interactions.onCameraClosed()
-        }
+        onResult = interactions::onCameraResult
     )
 
     LaunchedEffect(state.isCameraOpen) {
