+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/utils/CameraManager.kt
@@ -1,0 +1,12 @@
+package net.thechance.mena.core_chat.presentation.utils
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.ImageBitmap
+
+@Composable
+expect fun rememberCameraManager(onResult: (ImageBitmap?) -> Unit): CameraManager
+
+
+expect class CameraManager{
+    fun launch()
+}
