+++ core_chat_presentation/src/iosMain/kotlin/net/thechance/mena/core_chat/presentation/utils/CameraManager.ios.kt
@@ -1,20 +1,22 @@
-package net.thechance.mena.core_chat.presentation.camera
+package net.thechance.mena.core_chat.presentation.utils
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.graphics.ImageBitmap
 import io.github.vinceglb.filekit.dialogs.compose.PhotoResultLauncher
 import io.github.vinceglb.filekit.dialogs.compose.rememberCameraPickerLauncher
 import io.github.vinceglb.filekit.dialogs.compose.util.toImageBitmap
 import kotlinx.coroutines.launch
-import net.thechance.mena.core_chat.presentation.utils.encodeToByteArrayWithCompressionToMaxSize
 
 @Composable
-actual fun rememberCameraManager(onResult: (ByteArray?) -> Unit): CameraManager {
+actual fun rememberCameraManager(onResult: (ImageBitmap?) -> Unit): CameraManager {
     val scope = rememberCoroutineScope()
     val launcher = rememberCameraPickerLauncher { file ->
-        file?.let { image ->
+        if (file == null) {
+            onResult(null)
+        } else {
             scope.launch {
-                onResult(image.toImageBitmap().encodeToByteArrayWithCompressionToMaxSize())
+                onResult(file.toImageBitmap())
             }
         }
     }
