--- core_chat_presentation/src/main/res/xml/path_provider.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<paths>
-    <files-path
-        name="cache_images"
-        path="/" />
-    <cache-path
-        name="cache_images"
-        path="/" />
-</paths>
